==> 01. CorpusChecks.ipynb <==
## Checking Data

This was a general first check of the data gathered by each scraper separately, to make sure things worked correctly with the PushShift API. It is not very important in the context of the analysis.

import pandas as pd
import numpy as np
from textblob import TextBlob
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

casomc = pd.read_csv('../data/CasualConvOMC.csv')
happy = pd.read_csv('../data/happy_15k.csv')
depression = pd.read_csv('../data/depression_30k.csv')
anxiety = pd.read_csv('../data/anxiety_20k.csv')

casomc.head()

happy.shape

casomc.shape

analyzer = SentimentIntensityAnalyzer()

cc_sample = casomc[casomc['subreddit']=='CasualConversation'].sample(100)
analyzer.polarity_scores(cc_sample['title'])

omc_sample = casomc[casomc['subreddit']=='offmychest'].sample(100)
analyzer.polarity_scores(omc_sample['title'])

casomc['subreddit'].unique()

#### Creating a BaseText Corpus from CasualConversation and Happy

basetext = pd.concat([casomc[casomc['subreddit']=='CasualConversation'],happy])

basetext.shape

basetext.drop(columns = 'Unnamed: 0', inplace=True)

basetext.head()

#### Extracting a Sample and Seeing if We Can Attach Vader Polarity Scores to the Corpus

base_sample = basetext.sample(100)

res = analyzer.polarity_scores(base_sample['title'])

[res['neg'], res['neu']]

def sentiment(row):
    analyzer = SentimentIntensityAnalyzer()
    res = analyzer.polarity_scores(row)
    return pd.Series([res['neg'], res['neu'], res['pos'], res['compound']])

base_sample[['neg','neu','pos','comp']] = base_sample['title'].apply(sentiment)

base_sample.head()

#### Let's do a quick check on the sentiments of our base class, Depression class, and Anxiety class using sampling to see if our base class will be different enough

analyzer.polarity_scores(basetext.sample(500)['title'])

analyzer.polarity_scores(depression.sample(500)['title'])

analyzer.polarity_scores(anxiety.sample(500)['title'])

##### Looking good compared to intuition. Now we can export our base text corpus for use on the more robust classifier.

#### Exporting the basetext corpus

basetext.head()

basetext.to_csv('./basetext.csv')
