==> 02. Cleaning.ipynb <==
## Data Cleaning and Corpus Creation

Here we bring in all of the text gathered by our scrapers, clean all of it, and create a full corpus with 80k total posts: 30k from r/depression, 15k each from r/CasualConversation and r/happy, and 20k from r/Anxiety. Note that in later notebooks, the Anxiety section of the corpus is excluded and analysis is done only on Depression and Neutral Text.

import pandas as pd
import numpy as np

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction import text
import re
from nltk.tokenize import RegexpTokenizer, word_tokenize
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
pd.set_option('display.max_columns', None)

basetext = pd.read_csv('../data/basetext.csv').drop(columns='Unnamed: 0')
depression = pd.read_csv('../data/depression_30k.csv').drop(columns='Unnamed: 0')
anxiety = pd.read_csv('../data/anxiety_20k.csv').drop(columns='Unnamed: 0')

corpus = pd.concat([basetext,depression,anxiety])

corpus['subreddit'].unique()

corpus['class'] = corpus['subreddit'].map({'CasualConversation':0, 'happy':0, 'depression':1, 'Anxiety':2})

corpus.drop(columns=['author','created_utc'],inplace=True)
corpus = corpus.replace('[removed]', ' ').replace('[deleted]',' ')
corpus = corpus.fillna(' ')

corpus['full_text'] = corpus['title'] + " " + corpus['selftext']

corpus['full_text'] = corpus['full_text'].map(lambda x:x.lower())
corpus['full_text'] = corpus['full_text'].map(lambda x:x.replace("'ve",' have'))

corpus = corpus[['full_text','subreddit','class']]

my_tokenizer = RegexpTokenizer("[a-zA-Z0-9\'?]+")

my_stop = ['\n&gt;' ,'https', 'com', 'www', 'don','didn']
stop_words = text.ENGLISH_STOP_WORDS.union(my_stop)

cvect = CountVectorizer(stop_words=stop_words,max_features=100)

dtm_corp = cvect.fit_transform(corpus['full_text'])
dtm_corp = pd.DataFrame(dtm_corp.toarray(),columns=cvect.get_feature_names())

corp_top_60 = pd.DataFrame(dtm_corp.sum().nlargest(60))

### Splitting and Looking at Corpus by Class

dep = corpus[corpus['subreddit']=='depression']

dtm_dep = cvect.fit_transform(dep['full_text'])
dtm_dep = pd.DataFrame(dtm_dep.toarray(),columns=cvect.get_feature_names())

dep_top_60 = pd.DataFrame(dtm_dep.sum().nlargest(60))

anx = corpus[corpus['subreddit']=='Anxiety']

dtm_anx = cvect.fit_transform(anx['full_text'])
dtm_anx = pd.DataFrame(dtm_anx.toarray(),columns=cvect.get_feature_names())

anx_top_60 = pd.DataFrame(dtm_anx.sum().nlargest(60))

base = corpus[(corpus['subreddit']!='depression') & (corpus['subreddit']!='Anxiety')]

dtm_base = cvect.fit_transform(base['full_text'])
dtm_base = pd.DataFrame(dtm_base.toarray(),columns=cvect.get_feature_names())

base_top_60 = pd.DataFrame(dtm_base.sum().nlargest(60))

dep_top_60.T

anx_top_60.T

base_top_60.T

corpus

analyzer = SentimentIntensityAnalyzer()

def sentiment(row):
    analyzer = SentimentIntensityAnalyzer()
    res = analyzer.polarity_scores(row)
    return pd.Series([res['neg'], res['neu'], res['pos'], res['compound']])

corpus[['neg','neu','pos','comp']] = corpus['full_text'].apply(sentiment)

corpus

corpus.groupby('subreddit')['neg'].agg(np.mean)

corpus.to_csv('../data/corpus.csv')
