==> 03. DAICScratch.ipynb <==
## Investigating DAIC-WOZ Data

In the following, I conduced some general EDA/unpacking of selected datapoints from the DAIC Corpus

import pandas as pd

tr_300 = pd.read_csv('../data/300_TRANSCRIPT.csv',sep = '\\t',engine='python')

tr_300[tr_300['speaker']!='Ellie']
