==> 04. BERTVectorCreation-Copy1.ipynb <==
## BERT Vector Creation for Binary Depression Classifier

In the following notebook, the entire corpus consisting of depressive text and neutral text, 30,000 reddit posts each, are converted to bERT vectors via spaCy's transformer module and the distilbert language model, which offers faster processing with minimal losses in semantic accuracy.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
plt.style.use('ggplot')
import seaborn as sns

from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

import spacy
from sklearn.pipeline import make_pipeline
from sklearn.model_selection import train_test_split

from nltk.tokenize import RegexpTokenizer, word_tokenize
import re

import tensorflow as tf
from tensorflow.keras.models import Sequential, load_model
from tensorflow.keras.layers import Dense, Dropout, GRU, Input, Flatten, LSTM, Conv1D, MaxPooling1D, GlobalMaxPooling1D, Input, Embedding, Bidirectional
from tensorflow.keras.preprocessing.sequence import pad_sequences
from tensorflow.keras.wrappers.scikit_learn import KerasClassifier, KerasRegressor
from tensorflow.keras.callbacks import EarlyStopping
from tensorflow.keras.regularizers import l2

corpus = pd.read_csv('../data/corpus.csv').drop(columns='Unnamed: 0')

corpus

The full file includes posts from r/Anxiety, which we will not use for this analysis, so we exclude them to get depression and neutral text.

binary_df = corpus[corpus['subreddit']!='Anxiety']

binary_df.head()

### Getting Bert Vectors for the Whole Corpus

X = binary_df['full_text']
y = binary_df['subreddit']

X_train, X_test, y_train, y_test = train_test_split(X,y,test_size=0.2)

import numpy as np
import spacy
from sklearn.base import BaseEstimator, TransformerMixin

class WordVectorTransformer(TransformerMixin,BaseEstimator):
    def __init__(self, model="en_trf_distilbertbaseuncased_lg"):    #put bert embeddings here
        self.model = model
    def fit(self,X,y=None):
        return self
    def transform(self,X):
        nlp = spacy.load(self.model)
        return np.concatenate([nlp(doc).vector.reshape(1,-1) for doc in X])

bertvect = WordVectorTransformer()

X_train_bvect = bertvect.fit_transform(X_train)
X_test_bvect = bertvect.transform(X_test)

y_train_cat = y_train.map({'depression':1,'CasualConversation':0,'happy':0})
y_train_cat

y_test_cat = y_test.map({'depression':1,'CasualConversation':0,'happy':0})
y_test_cat

y_train_vect = tf.keras.utils.to_categorical(y_train_cat)
y_test_vect = tf.keras.utils.to_categorical(y_test_cat)

X_train_bvect.shape

X_test_bvect.shape

y_train_vect.shape

y_test_vect.shape

# save numpy array as npy file
from numpy import asarray
from numpy import save
from numpy import load

### Saving Vectors and Corresponding Classes as Numpy Arrays

save('X_train_bvect', X_train_bvect)

save('X_test_bvect', X_test_bvect)

save('y_train_vect', y_train_vect)

save('y_test_vect', y_test_vect)

peek = load('X_train_bvect.npy')

peek[:2]

X_train_bvect[:2]

X_train_reshape = X_train_bvect.reshape(-1,768,1)
X_test_reshape = X_test_bvect.reshape(-1,768,1)

X_train_reshape.shape

X_test_reshape.shape

### Testing the Vectors Out on a Neural Net Classifier, One Epoch

model_l = Sequential()

model_l.add(Conv1D(32, 7, activation = 'relu'))
model_l.add(MaxPooling1D())
model_l.add(Bidirectional(LSTM(24)))
model_l.add(Dense(64,activation='relu',kernel_regularizer=l2(0.001)))
model_l.add(Dropout(0.5))
model_l.add(Dense(64,activation='relu',kernel_regularizer=l2(0.001)))
model_l.add(Dropout(0.5))
model_l.add(Dense(1,activation='sigmoid'))

model_l.compile(optimizer='nadam', metrics=['accuracy'], loss='binary_crossentropy')

history_l = model_l.fit(X_train_reshape, y_train_cat.to_numpy(), validation_data=(X_test_reshape,y_test_cat.to_numpy()), epochs=1)

preds = model_l.predict(X_test_reshape[:10])
preds

save('y_test_cat',y_test_cat.to_numpy())

save('y_train_cat',y_train_cat.to_numpy())

X_train

y_train

X_train.to_csv('../data/X_train.csv')
X_test.to_csv('../data/X_test.csv')

y_train.to_csv('../data/y_train.csv')
y_test.to_csv('../data/y_test.csv')
