==> 09. SuicidalTextWarningSystem.ipynb <==
## Suicidal Text Warning System

The following sets up a simple warning system for tect input that will output a specific warning if the input contains certain words/phrases that may indicate intention of self-harm or suicide. It is a very rudimentary text-matching system, and will likely overpredict/overwarn.

A **trigger warning** is also in order for the following despite its brevity, as some of the included high-risk words and phrases may be difficult to read.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
plt.style.use('ggplot')
import seaborn as sns

from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

import spacy
from sklearn.pipeline import make_pipeline
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction import text
from sklearn.preprocessing import StandardScaler

corpus = pd.read_csv('../data/corpus.csv').drop(columns='Unnamed: 0')

warning_list = ['die', 'suicide', 'suicidal', 'death', 'dead', 'kill', 'end it', 'end my life', 'cut', 'hurt myself']

text_input = input('How are you feeling?')

for i in warning_list:
    if i in text_input:
        print('Your input may contain references to self-harm or suicide. If you are considering self-harm or suicide, please reach out to this helpline: https://www.crisistextline.org/topics/self-harm/#what-is-self-harm-1 or call the suicide hotline number at 800-273-8255.')
